Create README._for_Text_to_Speech.md
This is the README.md file for Digital Substitute for Braille. It contains all details regarding the features, and requirements of this TTS(Text-to-Speech) app. It consists of setup instructions and the application workflow. In addition to this, there are also troubleshooting tips to make sure that you are all set to run your app in a hassle-free manner.